+++ src/codroid/__about__.py
@@ -1,4 +1,4 @@
 # SPDX-FileCopyrightText: 2026-present guybod <b13140185898@outlook.com>
 #
 # SPDX-License-Identifier: MIT
-__version__ = "2.1.7"
+__version__ = "2.1.8"